[add]create base model
