+++ web_game/src/main/java/com/spring/web_game/model/GameModel.java
@@ -3,26 +3,16 @@ package com.spring.web_game.model;
 import java.util.HashSet;
 import java.util.Set;
 
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-public class GameModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
+public class GameModel extends BaseModel{
     private String gameTitle;
     private String gameCode;
     private boolean isHightestScoreBest;
     private Set<ScoreModel> scoreList = new HashSet<>();
 
-    public Long getId() {
-        return id;
+    GameModel(Long id){
+        super();
     }
-    public void setId(Long id) {
-        this.id = id;
-    }
+
     public String getGameTitle() {
         return gameTitle;
     }
@@ -46,30 +36,5 @@ public class GameModel {
     }
     public void setScoreList(Set<ScoreModel> scoreList) {
         this.scoreList = scoreList;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        GameModel other = (GameModel) obj;
-        if (id == null) {
-            if (other.id != null)
-                return false;
-        } else if (!id.equals(other.id))
-            return false;
-        return true;
     }
 }
