+++ web_game/src/main/java/com/spring/web_game/model/PlayerModel.java
@@ -3,16 +3,9 @@ package com.spring.web_game.model;
 import java.sql.Blob;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 
 @Entity
-public class PlayerModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
+public class PlayerModel extends BaseModel{
     private String firstName;
     private String familyName;
     private String emailAddr;
@@ -20,14 +13,7 @@ public class PlayerModel {
     private Blob profileImg;
 
     public PlayerModel(){
-
-    }
-
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
+        super();
     }
     public String getFirstName() {
         return firstName;
@@ -58,30 +44,5 @@ public class PlayerModel {
     }
     public void setProfileImg(Blob profileImg) {
         this.profileImg = profileImg;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        PlayerModel other = (PlayerModel) obj;
-        if (id == null) {
-            if (other.id != null)
-                return false;
-        } else if (!id.equals(other.id))
-            return false;
-        return true;
     }
 }
